Moved the npm clients to a dedicated space
