+++ src/tooling/piral-cli/src/npm-clients/yarn.ts
@@ -1,13 +1,13 @@
 import { resolve } from 'path';
-import { log } from '../log';
-import { findFile } from '../io';
-import { runCommand } from '../scripts';
-import { MemoryStream } from '../MemoryStream';
+import { log } from '../common/log';
+import { findFile } from '../common/io';
+import { runCommand } from '../common/scripts';
+import { MemoryStream } from '../common/MemoryStream';
 
 // Helpers:
 
 function runYarnProcess(args: Array<string>, target: string, output?: NodeJS.WritableStream) {
-  log('generalDebug_0003', 'Starting the Yarn PnP process ...');
+  log('generalDebug_0003', 'Starting the Yarn Classic process ...');
   const cwd = resolve(process.cwd(), target);
   return runCommand('yarn', args, cwd, output);
 }
@@ -33,26 +33,26 @@ function convert(flags: Array<string>) {
 export async function installDependencies(target = '.', ...flags: Array<string>) {
   const ms = new MemoryStream();
   await runYarnProcess(['install', ...convert(flags)], target, ms);
-  log('generalDebug_0003', `Yarn PnP install dependencies result: ${ms.value}`);
+  log('generalDebug_0003', `Yarn Classic install dependencies result: ${ms.value}`);
   return ms.value;
 }
 
 export async function uninstallPackage(packageRef: string, target = '.', ...flags: Array<string>) {
   const ms = new MemoryStream();
   await runYarnProcess(['remove', packageRef, ...convert(flags)], target, ms);
-  log('generalDebug_0003', `Yarn PnP remove package result: ${ms.value}`);
+  log('generalDebug_0003', `Yarn Classic remove package result: ${ms.value}`);
   return ms.value;
 }
 
 export async function installPackage(packageRef: string, target = '.', ...flags: Array<string>) {
   const ms = new MemoryStream();
   await runYarnProcess(['add', packageRef, ...convert(flags)], target, ms);
-  log('generalDebug_0003', `Yarn PnP add package result: ${ms.value}`);
+  log('generalDebug_0003', `Yarn Classic add package result: ${ms.value}`);
   return ms.value;
 }
 
 export async function detectClient(root: string, stopDir = resolve(root, '/')) {
-  return !!(await findFile(root, '.pnp.cjs', stopDir));
+  return !!(await findFile(root, 'yarn.lock', stopDir));
 }
 
 export async function initProject(projectName: string, target: string) {}
